fix: added comments fixed directory structure
