+++ secretmanager/snippets/regional_samples/create_regional_secret.py
@@ -54,7 +54,7 @@ def create_regional_secret(
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
-
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
     
     # Create the Secret Manager client.
