+++ secretmanager/snippets/regional_samples/delete_regional_secret.py
@@ -13,38 +13,38 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for listing secret versions of a
-regional secret.
+command line application and sample code for deleting an existing regional 
+secret.
 """
 
 import argparse
 
 
-# [START secretmanager_v1_list_regional_secret_versions]
-def list_regional_secret_versions(project_id: str, location_id: str, secret_id: str) -> None:
+# [START secretmanager_v1_delete_regional_secret]
+def delete_regional_secret(project_id: str, location_id: str, secret_id: str) -> None:
     """
-    List all secret versions in the given secret and their metadata.
+    Delete the secret with the given name and all of its versions.
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
     client = secretmanager_v1.SecretManagerServiceClient(client_options={
         "api_endpoint": api_endpoint
             })
-    
-    # Build the resource name of the parent secret.
-    parent = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
 
-    # List all secret versions.
-    for version in client.list_secret_versions(request={"parent": parent}):
-        print(f"Found secret version: {version.name}")
+    # Build the resource name of the secret.
+    name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
+
+    # Delete the secret.
+    client.delete_secret(request={"name": name})
 
 
-# [END secretmanager_v1_list_regional_secret_versions]
+# [END secretmanager_v1_delete_regional_secret]
 
 
 if __name__ == "__main__":
@@ -53,7 +53,7 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("secret_id", help="id of the secret in which to list")
+    parser.add_argument("secret_id", help="id of the secret to delete")
     args = parser.parse_args()
 
-    list_secret_versions(args.project_id, args.location_id, args.secret_id)
+    delete_secret(args.project_id, args.location_id, args.secret_id)
