+++ secretmanager/snippets/regional_samples/delete_regional_secret_with_etag.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright 2020 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,26 +13,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for granting access to a regional
+command line application and sample code for deleting an existing regional
 secret.
 """
 
 import argparse
 
-from google.iam.v1 import iam_policy_pb2  # type: ignore
 
-
-# [START secretmanager_v1_iam_grant_access_regional_secret]
-def iam_grant_access_with_regional_secret(
-    project_id: str, location_id: str, secret_id: str, member: str
-) -> iam_policy_pb2.SetIamPolicyRequest:
+# [START secretmanager_v1_delete_regional_secret_with_etag]
+def delete_regional_secret_with_etag(project_id: str,location_id: str, secret_id: str, etag: str) -> None:
     """
-    Grant the given member access to a secret.
+    Delete the regional secret with the given name, etag, and all of its versions.
     """
 
-    # Import the Secret Manager client library.
+    # Import the Secret Manager client library and types.
     from google.cloud import secretmanager_v1
+    from google.cloud.secretmanager_v1.types import service
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -43,20 +41,16 @@ def iam_grant_access_with_regional_secret(
     # Build the resource name of the secret.
     name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
 
-    # Get the current IAM policy.
-    policy = client.get_iam_policy(request={"resource": name})
+    # Build the request
+    request = service.DeleteSecretRequest()
+    request.name = name
+    request.etag = etag
 
-    # Add the given member with access permissions.
-    policy.bindings.add(role="roles/secretmanager.secretAccessor", members=[member])
+    # Delete the secret.
+    client.delete_secret(request=request)
 
-    # Update the IAM Policy.
-    new_policy = client.set_iam_policy(request={"resource": name, "policy": policy})
 
-    # Print data about the secret.
-    print(f"Updated IAM policy on {secret_id}")
-    # [END secretmanager_v1_iam_grant_access_regional_secret]
-
-    return new_policy
+# [END secretmanager_v1_delete_regional_secret_with_etag]
 
 
 if __name__ == "__main__":
@@ -65,8 +59,8 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("secret_id", help="id of the secret to get")
-    parser.add_argument("member", help="member to grant access")
+    parser.add_argument("secret_id", help="id of the secret to delete")
+    parser.add_argument("etag", help="current etag of the secret to delete")
     args = parser.parse_args()
 
-    iam_grant_access(args.project_id, args.location_id, args.secret_id, args.member)
+    delete_secret_with_etag(args.project_id, args.location_id, args.secret_id, args.etag)
