+++ secretmanager/snippets/regional_samples/destroy_regional_secret_version_with_etag.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright 2019 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,8 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for destroying a regional 
-secret version.
+command line application and sample code for destroying a secret version.
 """
 
 import argparse
@@ -22,9 +21,9 @@ import argparse
 from google.cloud import secretmanager_v1
 
 
-# [START secretmanager_v1_destroy_regional_secret_version]
-def destroy_regional_secret_version(
-    project_id: str, location_id: str, secret_id: str, version_id: str
+# [START secretmanager_v1_destroy_regional_secret_version_with_etag]
+def destroy_regional_secret_version_with_etag(
+    project_id: str, location_id: str, secret_id: str, version_id: str, etag: str
 ) -> secretmanager_v1.DestroySecretVersionRequest:
     """
     Destroy the given secret version, making the payload irrecoverable. Other
@@ -33,7 +32,9 @@ def destroy_regional_secret_version(
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
+    from google.cloud.secretmanager_v1.types import service
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -44,11 +45,16 @@ def destroy_regional_secret_version(
     # Build the resource name of the secret version
     name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}/versions/{version_id}"
 
+    # Build the request
+    request = service.DestroySecretVersionRequest()
+    request.name = name
+    request.etag = etag
+
     # Destroy the secret version.
-    response = client.destroy_secret_version(request={"name": name})
+    response = client.destroy_secret_version(request=request)
 
     print(f"Destroyed secret version: {response.name}")
-    # [END secretmanager_v1_destroy_regional_secret_version]
+    # [END secretmanager_v1_destroy_regional_secret_version_with_etag]
 
     return response
 
@@ -61,6 +67,9 @@ if __name__ == "__main__":
     parser.add_argument("location_id", help="id of location where secret is stored")
     parser.add_argument("secret_id", help="id of the secret from which to act")
     parser.add_argument("version_id", help="id of the version to destroy")
+    parser.add_argument("etag", help="current etag of the version")
     args = parser.parse_args()
 
-    destroy_secret_version(args.project_id, args.location_id, args.secret_id, args.version_id)
+    destroy_secret_version_with_etag(
+        args.project_id, args.location_id, args.secret_id, args.version_id, args.etag
+    )
