+++ secretmanager/snippets/regional_samples/enable_regional_secret_version.py
@@ -13,37 +13,45 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for deleting an existing regional 
-secret.
+command line application and sample code for enabling a regional
+secret version.
 """
 
 import argparse
 
+from google.cloud import secretmanager_v1
 
-# [START secretmanager_v1_delete_regional_secret]
-def delete_regional_secret(project_id: str, location_id: str, secret_id: str) -> None:
+
+# [START secretmanager_enable_secret_version]
+def enable_regional_secret_version(
+    project_id: str, location_id: str, secret_id: str, version_id: str
+) -> secretmanager_v1.EnableSecretVersionRequest:
     """
-    Delete the secret with the given name and all of its versions.
+    Enable the given secret version, enabling it to be accessed after
+    previously being disabled. Other secrets versions are unaffected.
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
     client = secretmanager_v1.SecretManagerServiceClient(client_options={
         "api_endpoint": api_endpoint
             })
+    
+    # Build the resource name of the secret version
+    name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}/versions/{version_id}"
 
-    # Build the resource name of the secret.
-    name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
+    # Disable the secret version.
+    response = client.enable_secret_version(request={"name": name})
 
-    # Delete the secret.
-    client.delete_secret(request={"name": name})
+    print(f"Enabled secret version: {response.name}")
+    # [END secretmanager_enable_secret_version]
 
-
-# [END secretmanager_v1_delete_regional_secret]
+    return response
 
 
 if __name__ == "__main__":
@@ -52,7 +60,8 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("secret_id", help="id of the secret to delete")
+    parser.add_argument("secret_id", help="id of the secret from which to act")
+    parser.add_argument("version_id", help="id of the version to enable")
     args = parser.parse_args()
 
-    delete_secret(args.project_id, args.location_id, args.secret_id)
+    enable_secret_version(args.project_id, args.location_id, args.secret_id, args.version_id)
