+++ secretmanager/snippets/regional_samples/get_regional_secret_version.py
@@ -13,24 +13,28 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for accessing a regional secret version.
+command line application and sample code for getting metadata about a regional 
+secret version, but not the secret payload.
 """
 
 import argparse
 
-# [START secretmanager_access_secret_version]
 from google.cloud import secretmanager_v1
-import google_crc32c
 
 
-def access_regional_secret_version(
+# [START secretmanager_v1_get_regional_secret_version]
+def get_regional_secret_version(
     project_id: str, location_id: str, secret_id: str, version_id: str
-) -> secretmanager_v1.AccessSecretVersionResponse:
+    ) -> secretmanager_v1.GetSecretVersionRequest:
     """
-    Access the payload for the given secret version if one exists. The version
-    can be a version number as a string (e.g. "5") or an alias (e.g. "latest").
+    Get information about the given secret version. It does not include the
+    payload data.
     """
 
+    # Import the Secret Manager client library.
+    from google.cloud import secretmanager_v1
+
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -41,23 +45,13 @@ def access_regional_secret_version(
     # Build the resource name of the secret version.
     name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}/versions/{version_id}"
 
-    # Access the secret version.
-    response = client.access_secret_version(request={"name": name})
+    # Get the secret version.
+    response = client.get_secret_version(request={"name": name})
 
-    # Verify payload checksum.
-    crc32c = google_crc32c.Checksum()
-    crc32c.update(response.payload.data)
-    if response.payload.data_crc32c != int(crc32c.hexdigest(), 16):
-        print("Data corruption detected.")
-        return response
-
-    # Print the secret payload.
-    #
-    # WARNING: Do not print the secret in a production environment - this
-    # snippet is showing how to access the secret material.
-    payload = response.payload.data.decode("UTF-8")
-    print(f"Plaintext: {payload}")
-    # [END secretmanager_access_secret_version]
+    # Print information about the secret version.
+    state = response.state.name
+    print(f"Got secret version {response.name} with state {state}")
+    # [END secretmanager_v1_get_regional_secret_version]
 
     return response
 
@@ -68,8 +62,8 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("secret_id", help="id of the secret to access")
-    parser.add_argument("version_id", help="version to access")
+    parser.add_argument("secret_id", help="id of the secret from which to act")
+    parser.add_argument("version_id", help="id of the version to get")
     args = parser.parse_args()
 
-    access_secret_version(args.project_id, args.location_id, args.secret_id, args.version_id)
+    get_secret_version(args.project_id, args.location_id, args.secret_id, args.version_id)
