+++ secretmanager/snippets/regional_samples/iam_revoke_access_with_regional_secret.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright 2019 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,24 +13,27 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for getting metadata about a regional secret.
+command line application and sample code for revoking access to a regional 
+secret.
 """
 
 import argparse
 
-from google.cloud import secretmanager_v1
+from google.iam.v1 import iam_policy_pb2  # type: ignore
 
 
-# [START secretmanager_v1_get_regional_secret]
-def get_regional_secret(project_id: str, location_id: str, secret_id: str) -> secretmanager_v1.GetSecretRequest:
+# [START secretmanager_v1_iam_revoke_access_regional_secret]
+def iam_revoke_access_with_regional_secret(
+    project_id: str, location_id: str, secret_id: str, member: str
+) -> iam_policy_pb2.SetIamPolicyRequest:
     """
-    Get information about the given secret. This only returns metadata about
-    the secret container, not any secret material.
+    Revoke the given member access to a secret.
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -41,14 +44,23 @@ def get_regional_secret(project_id: str, location_id: str, secret_id: str) -> se
     # Build the resource name of the secret.
     name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
 
-    # Get the secret.
-    response = client.get_secret(request={"name": name})
+    # Get the current IAM policy.
+    policy = client.get_iam_policy(request={"resource": name})
+
+    # Remove the given member's access permissions.
+    accessRole = "roles/secretmanager.secretAccessor"
+    for b in list(policy.bindings):
+        if b.role == accessRole and member in b.members:
+            b.members.remove(member)
+
+    # Update the IAM Policy.
+    new_policy = client.set_iam_policy(request={"resource": name, "policy": policy})
 
     # Print data about the secret.
-    print(f"Got secret {response.name}")
-    # [END secretmanager_v1_get_regional_secret]
+    print(f"Updated IAM policy on {secret_id}")
+    # [END secretmanager_v1_iam_revoke_access_regional_secret]
 
-    return response
+    return new_policy
 
 
 if __name__ == "__main__":
@@ -58,6 +70,7 @@ if __name__ == "__main__":
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
     parser.add_argument("secret_id", help="id of the secret to get")
+    parser.add_argument("member", help="member to revoke access")
     args = parser.parse_args()
 
-    get_secret(args.project_id, args.location_id, args.secret_id)
+    iam_revoke_access(args.project_id, args.location_id, args.secret_id, args.member)
