+++ secretmanager/snippets/regional_samples/list_regional_secret_versions.py
@@ -13,44 +13,39 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for disabling a regional 
-secret version.
+command line application and sample code for listing secret versions of a
+regional secret.
 """
 
 import argparse
 
-from google.cloud import secretmanager_v1
 
-
-# [START secretmanager_v1_disable_regional_secret_version]
-def disable_regional_secret_version(
-    project_id: str, location_id: str, secret_id: str, version_id: str
-) -> secretmanager_v1.DisableSecretVersionRequest:
+# [START secretmanager_v1_list_regional_secret_versions]
+def list_regional_secret_versions(project_id: str, location_id: str, secret_id: str) -> None:
     """
-    Disable the given secret version. Future requests will throw an error until
-    the secret version is enabled. Other secrets versions are unaffected.
+    List all secret versions in the given secret and their metadata.
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
     client = secretmanager_v1.SecretManagerServiceClient(client_options={
         "api_endpoint": api_endpoint
             })
+    
+    # Build the resource name of the parent secret.
+    parent = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
 
-    # Build the resource name of the secret version
-    name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}/versions/{version_id}"
+    # List all secret versions.
+    for version in client.list_secret_versions(request={"parent": parent}):
+        print(f"Found secret version: {version.name}")
 
-    # Disable the secret version.
-    response = client.disable_secret_version(request={"name": name})
 
-    print(f"Disabled secret version: {response.name}")
-    # [END secretmanager_v1_disable_regional_secret_version]
-
-    return response
+# [END secretmanager_v1_list_regional_secret_versions]
 
 
 if __name__ == "__main__":
@@ -59,8 +54,7 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("secret_id", help="id of the secret from which to act")
-    parser.add_argument("version_id", help="id of the version to disable")
+    parser.add_argument("secret_id", help="id of the secret in which to list")
     args = parser.parse_args()
 
-    disable_secret_version(args.project_id, args.location_id, args.secret_id, args.version_id)
+    list_secret_versions(args.project_id, args.location_id, args.secret_id)
