+++ secretmanager/snippets/regional_samples/list_regional_secret_versions_with_filter.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright 2019 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,27 +13,29 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 """
-command line application and sample code for enabling a regional
-secret version.
+command line application and sample code for listing secret versions of a
+regional secret.
 """
 
-import argparse
 
-from google.cloud import secretmanager_v1
-
-
-# [START secretmanager_enable_secret_version]
-def enable_regional_secret_version(
-    project_id: str, location_id: str, secret_id: str, version_id: str
-) -> secretmanager_v1.EnableSecretVersionRequest:
+# [START secretmanager_v1_list_regional_secret_versions_with_filter]
+def list_regional_secret_versions_with_filter(
+    project_id: str, location_id: str, secret_id: str, filter_str: str = "state:ENABLED"
+) -> None:
     """
-    Enable the given secret version, enabling it to be accessed after
-    previously being disabled. Other secrets versions are unaffected.
+    List all secret versions in the given secret and their metadata.
+
+    Args:
+      project_id: Parent project id
+      secret_id: Parent secret id
+      filter_str: Secret version filter, constructing according to
+                  https://cloud.google.com/secret-manager/docs/filtering
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -41,18 +43,17 @@ def enable_regional_secret_version(
         "api_endpoint": api_endpoint
             })
     
-    # Build the resource name of the secret version
-    name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}/versions/{version_id}"
+    # Build the resource name of the parent secret.
+    parent = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
 
-    # Disable the secret version.
-    response = client.enable_secret_version(request={"name": name})
-
-    print(f"Enabled secret version: {response.name}")
-    # [END secretmanager_enable_secret_version]
-
-    return response
+    # List all secret versions.
+    for version in client.list_secret_versions(
+        request={"parent": parent, "filter": filter_str}
+    ):
+        print(f"Found secret version: {version.name}")
 
 
+    # [END secretmanager_v1_list_regional_secret_versions_with_filter]
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description=__doc__, formatter_class=argparse.RawDescriptionHelpFormatter
@@ -60,7 +61,7 @@ if __name__ == "__main__":
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
     parser.add_argument("secret_id", help="id of the secret from which to act")
-    parser.add_argument("version_id", help="id of the version to enable")
+    parser.add_argument("filter", help="filter string to apply")
     args = parser.parse_args()
 
-    enable_secret_version(args.project_id, args.location_id, args.secret_id, args.version_id)
+    list_secret_versions_with_filter(args.project_id, args.secret_id, args.filter)
