+++ secretmanager/snippets/regional_samples/list_regional_secrets.py
@@ -12,21 +12,24 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
+"""
+command line application and sample code for listing regional secrets 
+in a project.
+"""
 
 import argparse
 
-from google.cloud import secretmanager_v1
 
-
-# [START secretmanager_v1_update_regional_secret]
-def update_regional_secret(project_id: str, location_id: str, secret_id: str) -> secretmanager_v1.UpdateSecretRequest:
+# [START secretmanager_v1_list_regional_secrets]
+def list_regional_secrets(project_id: str, location_id: str) -> None:
     """
-    Update the metadata about an existing secret.
+    List all regional secrets in the given project.
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -34,21 +37,15 @@ def update_regional_secret(project_id: str, location_id: str, secret_id: str) ->
         "api_endpoint": api_endpoint
             })
 
-    # Build the resource name of the secret.
-    name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
+    # Build the resource name of the parent project.
+    parent = f"projects/{project_id}/locations/{location_id}"
 
-    # Update the secret.
-    secret = {"name": name, "labels": {"secretmanager": "rocks"}}
-    update_mask = {"paths": ["labels"]}
-    response = client.update_secret(
-        request={"secret": secret, "update_mask": update_mask}
-    )
+    # List all secrets.
+    for secret in client.list_secrets(request={"parent": parent}):
+        print(f"Found secret: {secret.name}")
 
-    # Print the new secret name.
-    print(f"Updated secret: {response.name}")
-    # [END secretmanager_v1_update_regional_secret]
 
-    return response
+# [END secretmanager_v1_list_regional_secrets]
 
 
 if __name__ == "__main__":
@@ -57,7 +54,6 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("--secret-id", required=True)
     args = parser.parse_args()
 
-    update_secret(args.project_id, args.location_id, args.secret_id)
+    list_secrets(args.project_id, args.location_id)
