+++ secretmanager/snippets/regional_samples/list_regional_secrets_with_filter.py
@@ -32,6 +32,7 @@ def list_regional_secrets_with_filter(project_id: str, location_id: str,  filter
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
