+++ secretmanager/snippets/regional_samples/update_regional_secret_with_alias.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright 2020 Google LLC
+# Copyright 2022 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,27 +12,24 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
-"""
-command line application and sample code for revoking access to a regional 
-secret.
-"""
 
 import argparse
 
-from google.iam.v1 import iam_policy_pb2  # type: ignore
+from google.cloud import secretmanager_v1
 
 
-# [START secretmanager_v1_iam_revoke_access_regional_secret]
-def iam_revoke_access_with_regional_secret(
-    project_id: str, location_id: str, secret_id: str, member: str
-) -> iam_policy_pb2.SetIamPolicyRequest:
+# [START secretmanager_v1_update_regional_secret_with_alias]
+def update_regional_secret_with_alias(
+    project_id: str, location_id: str, secret_id: str
+) -> secretmanager_v1.UpdateSecretRequest:
     """
-    Revoke the given member access to a secret.
+    Update the metadata about an existing regional secret.
     """
 
     # Import the Secret Manager client library.
     from google.cloud import secretmanager_v1
 
+    # Endpoint to call the regional secret manager sever
     api_endpoint = f"secretmanager.{location_id}.rep.googleapis.com"
 
     # Create the Secret Manager client.
@@ -43,23 +40,18 @@ def iam_revoke_access_with_regional_secret(
     # Build the resource name of the secret.
     name = f"projects/{project_id}/locations/{location_id}/secrets/{secret_id}"
 
-    # Get the current IAM policy.
-    policy = client.get_iam_policy(request={"resource": name})
+    # Update the secret.
+    secret = {"name": name, "version_aliases": {"test": 1}}
+    update_mask = {"paths": ["version_aliases"]}
+    response = client.update_secret(
+        request={"secret": secret, "update_mask": update_mask}
+    )
 
-    # Remove the given member's access permissions.
-    accessRole = "roles/secretmanager.secretAccessor"
-    for b in list(policy.bindings):
-        if b.role == accessRole and member in b.members:
-            b.members.remove(member)
+    # Print the new secret name.
+    print(f"Updated secret: {response.name}")
+    # [END secretmanager_v1_update_regional_secret_with_alias]
 
-    # Update the IAM Policy.
-    new_policy = client.set_iam_policy(request={"resource": name, "policy": policy})
-
-    # Print data about the secret.
-    print(f"Updated IAM policy on {secret_id}")
-    # [END secretmanager_v1_iam_revoke_access_regional_secret]
-
-    return new_policy
+    return response
 
 
 if __name__ == "__main__":
@@ -68,8 +60,7 @@ if __name__ == "__main__":
     )
     parser.add_argument("project_id", help="id of the GCP project")
     parser.add_argument("location_id", help="id of location where secret is stored")
-    parser.add_argument("secret_id", help="id of the secret to get")
-    parser.add_argument("member", help="member to revoke access")
+    parser.add_argument("--secret-id", required=True)
     args = parser.parse_args()
 
-    iam_revoke_access(args.project_id, args.location_id, args.secret_id, args.member)
+    update_secret_with_alias(args.project_id, args.location_id, args.secret_id)
