+++ secretmanager/snippets/snippets_test.py
@@ -47,29 +47,29 @@ from update_secret import update_secret
 from update_secret_with_alias import update_secret_with_alias
 from update_secret_with_etag import update_secret_with_etag
 
-from access_regional_secret_version import access_regional_secret_version
-from add_regional_secret_version import add_regional_secret_version
-from create_regional_secret import create_regional_secret
-from delete_regional_secret import delete_regional_secret
-from delete_regional_secret_with_etag import delete_regional_secret_with_etag
-from destroy_regional_secret_version import destroy_regional_secret_version
-from destroy_regional_secret_version_with_etag import destroy_regional_secret_version_with_etag
-from disable_regional_secret_version import disable_regional_secret_version
-from disable_regional_secret_version_with_etag import disable_regional_secret_version_with_etag
-from enable_regional_secret_version import enable_regional_secret_version
-from enable_regional_secret_version_with_etag import enable_regional_secret_version_with_etag
-from get_regional_secret import get_regional_secret
-from get_regional_secret_version import get_regional_secret_version
-from iam_grant_access_with_regional_secret import iam_grant_access_with_regional_secret
-from iam_revoke_access_with_regional_secret import iam_revoke_access_with_regional_secret
-from list_regional_secret_versions import list_regional_secret_versions
-from list_regional_secret_versions_with_filter import list_regional_secret_versions_with_filter
-from list_regional_secrets import list_regional_secrets
-from list_regional_secrets_with_filter import list_regional_secrets_with_filter
-from update_regional_secret import update_regional_secret
-from update_regional_secret_with_alias import update_regional_secret_with_alias
-from update_regional_secret_with_etag import update_regional_secret_with_etag
-from regional_quickstart import regional_quickstart
+from regional_samples.access_regional_secret_version import access_regional_secret_version
+from regional_samples.add_regional_secret_version import add_regional_secret_version
+from regional_samples.create_regional_secret import create_regional_secret
+from regional_samples.delete_regional_secret import delete_regional_secret
+from regional_samples.delete_regional_secret_with_etag import delete_regional_secret_with_etag
+from regional_samples.destroy_regional_secret_version import destroy_regional_secret_version
+from regional_samples.destroy_regional_secret_version_with_etag import destroy_regional_secret_version_with_etag
+from regional_samples.disable_regional_secret_version import disable_regional_secret_version
+from regional_samples.disable_regional_secret_version_with_etag import disable_regional_secret_version_with_etag
+from regional_samples.enable_regional_secret_version import enable_regional_secret_version
+from regional_samples.enable_regional_secret_version_with_etag import enable_regional_secret_version_with_etag
+from regional_samples.get_regional_secret import get_regional_secret
+from regional_samples.get_regional_secret_version import get_regional_secret_version
+from regional_samples.iam_grant_access_with_regional_secret import iam_grant_access_with_regional_secret
+from regional_samples.iam_revoke_access_with_regional_secret import iam_revoke_access_with_regional_secret
+from regional_samples.list_regional_secret_versions import list_regional_secret_versions
+from regional_samples.list_regional_secret_versions_with_filter import list_regional_secret_versions_with_filter
+from regional_samples.list_regional_secrets import list_regional_secrets
+from regional_samples.list_regional_secrets_with_filter import list_regional_secrets_with_filter
+from regional_samples.update_regional_secret import update_regional_secret
+from regional_samples.update_regional_secret_with_alias import update_regional_secret_with_alias
+from regional_samples.update_regional_secret_with_etag import update_regional_secret_with_etag
+from regional_samples.regional_quickstart import regional_quickstart
 
 @pytest.fixture()
 def client() -> secretmanager.SecretManagerServiceClient:
